Split LogEntry
