+++ LokiPackage/Sources/Features/Debug/Log/LogScreen.swift
@@ -186,17 +186,6 @@ struct LogScreen: View {
     }
 }
 
-struct LogEntry: Sendable {
-    let message: String
-    let date: Date
-    let library: String
-    let processIdentifier: String
-    let threadIdentifier: String
-    let category: String
-    let subsystem: String
-    let level: OSLogEntryLog.Level
-}
-
 actor LogStore {
     func entries() throws -> [LogEntry] {
         let store = try OSLogStore(scope: .currentProcessIdentifier)
